References Oppevara/oppevara#42, initial version.
The code is almost fully functional, yet it will have to be rewritten a
bit before going live. The functionality will remain the same and it is
compatible with the library module (library might still need some
additional configurations added, this way all the field names would not
be hard-coded).